one files change missing yes! * reinterated trunk * made the engine side C interface compileable with a C compiler * added some things to easy up life for non native AI Interface libraries * NullAI is notw a real C AI
git-svn-id: https://spring.clan-sy.com/svn/spring/branches/caiinterface@6798 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Lua/LuaShaders.cpp
@@ -94,7 +94,7 @@ inline void CheckDrawingEnabled(lua_State* L, const char* caller)
 
 GLuint LuaShaders::GetProgramName(unsigned int progID) const
 {
-	if ((progID <= 0) || (progID >= (int)programs.size())) {
+	if ((progID <= 0) || (progID >= programs.size())) {
 		return 0;
 	}
 	return programs[progID].id;
@@ -126,7 +126,7 @@ unsigned int LuaShaders::AddProgram(const Program& p)
 
 void LuaShaders::RemoveProgram(unsigned int progID)
 {
-	if ((progID <= 0) || (progID >= (int)programs.size())) {
+	if ((progID <= 0) || (progID >= programs.size())) {
 		return;
 	}
 	Program& p = programs[progID];
